+++ aws-apigatewayv2/lib/standard-web-socket-api.ts
@@ -85,7 +85,7 @@ export class StandardWebSocketApi extends ExtendedConstruct {
       if (props.recordWeight) {
         domainName.createWeightedARecord(
           props.recordWeight,
-          props.evaluateTargetHealth ?? true
+          props.evaluateTargetHealth ?? true,
         );
       } else if (props.recordLatency) {
         domainName.createLatencyARecord(true);