+++ aws-cloudfront/lib/behavior-builder.ts
@@ -48,12 +48,12 @@ export class BehaviorBuilder extends ExtendedConstruct {
   constructor(
     scope: DistributionBuilder,
     origin: IOrigin,
-    path: string | undefined
+    path: string | undefined,
   ) {
     super(
       scope,
       (path === undefined || path === '' ? 'Default' : sha1sum(path)) +
-        'Behavior'
+        'Behavior',
     );
     this.path = path;
     scope.addBehavior(this, path);
@@ -69,7 +69,7 @@ export class BehaviorBuilder extends ExtendedConstruct {
 
   fallbackOrigin(
     fallbackOrigin?: IOrigin,
-    fallbackStatusCodes?: number[]
+    fallbackStatusCodes?: number[],
   ): BehaviorBuilder {
     if (fallbackOrigin) {
       // We need to reuse OriginGroups or it will complain about duplicate origins
@@ -133,7 +133,7 @@ export class BehaviorBuilder extends ExtendedConstruct {
   }
 
   originRequestPolicy(
-    originRequestPolicy?: IOriginRequestPolicy
+    originRequestPolicy?: IOriginRequestPolicy,
   ): BehaviorBuilder {
     this.options = {
       ...this.options,
@@ -143,7 +143,7 @@ export class BehaviorBuilder extends ExtendedConstruct {
   }
 
   responseHeadersPolicy(
-    responseHeadersPolicy?: IResponseHeadersPolicy
+    responseHeadersPolicy?: IResponseHeadersPolicy,
   ): BehaviorBuilder {
     this.options = {
       ...this.options,
@@ -161,7 +161,7 @@ export class BehaviorBuilder extends ExtendedConstruct {
   }
 
   viewerProtocolPolicy(
-    viewerProtocolPolicy?: ViewerProtocolPolicy
+    viewerProtocolPolicy?: ViewerProtocolPolicy,
   ): BehaviorBuilder {
     this.options = {
       ...this.options,
@@ -171,7 +171,7 @@ export class BehaviorBuilder extends ExtendedConstruct {
   }
 
   functionAssociations(
-    functionAssociations?: FunctionAssociation[]
+    functionAssociations?: FunctionAssociation[],
   ): BehaviorBuilder {
     this.options = {
       ...this.options,
@@ -198,7 +198,7 @@ export class BehaviorBuilder extends ExtendedConstruct {
     const redirectFunction = new RedirectFunction(
       this,
       'RedirectFunction',
-      props
+      props,
     );
     this.viewerRequestFunction(redirectFunction);
     return this;
@@ -242,12 +242,12 @@ export class BehaviorBuilder extends ExtendedConstruct {
     const cachePolicy = new StandardApiCachePolicy(
       this,
       'ApiCachePolicy',
-      additionalHeaders
+      additionalHeaders,
     );
     const originRequestPolicy = new StandardApiOriginRequestPolicy(
       this,
       'ApiOriginRequestPolicy',
-      additionalHeaders
+      additionalHeaders,
     );
     return this.allowedMethods(AllowedMethods.ALLOW_ALL)
       .cachePolicy(cachePolicy)
@@ -296,14 +296,14 @@ export class BehaviorBuilder extends ExtendedConstruct {
 
   behaviorFromCloudFromBucket(
     bucket: CloudFrontBucket,
-    path: string
+    path: string,
   ): BehaviorBuilder {
     return this.scope.behaviorFromCloudFromBucket(bucket, path);
   }
 
   behaviorFromDomainName(
     domainName: string | DomainName,
-    path: string
+    path: string,
   ): BehaviorBuilder {
     return this.scope.behaviorFromDomainName(domainName, path);
   }