+++ eslint.config.mjs
@@ -1,0 +1,22 @@
+// @ts-check
+
+import eslint from '@eslint/js';
+import tseslint from 'typescript-eslint';
+
+const config = [
+  {
+    ignores: [
+      'cdk.out',
+      'node_modules',
+      '**/*.js',
+      '**/*.d.*',
+      '**/*.map',
+      'test-files',
+    ],
+  },
+  ...tseslint.config(
+    eslint.configs.recommended,
+    ...tseslint.configs.recommended,
+  ),
+];
+export default config;