+++ CHANGELOG.md
@@ -7,10 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## v1.16.0
 
-### Changes
+### Changed
 
 -   Changes hashing algorithm used for identifiers on BehaviorBuilder.
 -   Dependency updates
+-   Replaced GTS with standard prettier and eslint configuration
 
 ## v1.15.2